Improve reusability of widget telemetry events.
