+++ graylog2-web-interface/src/views/components/widgets/Widget.tsx
@@ -18,7 +18,6 @@ import * as React from 'react';
 import { useCallback, useContext, useMemo, useState } from 'react';
 import styled from 'styled-components';
 
-import { getPathnameWithoutId } from 'util/URLUtils';
 import type { BackendWidgetPosition, WidgetResults, GetState } from 'views/types';
 import { widgetDefinition } from 'views/logic/Widgets';
 import type WidgetPosition from 'views/logic/widgets/WidgetPosition';
@@ -36,15 +35,17 @@ import useViewsDispatch from 'views/stores/useViewsDispatch';
 import { updateWidget, updateWidgetConfig } from 'views/logic/slices/widgetActions';
 import { selectActiveQuery } from 'views/logic/slices/viewSelectors';
 import { setTitle } from 'views/logic/slices/titlesActions';
-import useSendTelemetry from 'logic/telemetry/useSendTelemetry';
-import useLocation from 'routing/useLocation';
-import { TELEMETRY_EVENT_TYPE } from 'logic/telemetry/Constants';
 import useAutoRefresh from 'views/hooks/useAutoRefresh';
 import useViewType from 'views/hooks/useViewType';
 import View from 'views/logic/views/View';
 import IfDashboard from 'views/components/dashboard/IfDashboard';
 import FullSizeContainer from 'views/components/aggregationbuilder/FullSizeContainer';
 import type WidgetType from 'views/logic/widgets/Widget';
+import {
+  useSendWidgetEditTelemetry,
+  useSendWidgetEditCancelTelemetry,
+  useSendWidgetConfigUpdateTelemetry,
+} from 'views/components/widgets/telemety';
 
 import WidgetFrame from './WidgetFrame';
 import WidgetHeader from './WidgetHeader';
@@ -224,52 +225,40 @@ const Widget = ({ id, editing = false, widget, title, position, onPositionsChang
   const [oldWidget, setOldWidget] = useState(editing ? widget : undefined);
   const { focusedWidget, setWidgetEditing, unsetWidgetEditing } = useContext(WidgetFocusContext);
   const dispatch = useViewsDispatch();
-  const sendTelemetry = useSendTelemetry();
-  const { pathname } = useLocation();
+  const sendWidgetEditTelemetry = useSendWidgetEditTelemetry();
+  const sendWidgetEditCancelTelemetry = useSendWidgetEditCancelTelemetry();
+  const sendWidgetConfigUpdateTelemetry = useSendWidgetConfigUpdateTelemetry();
 
   const isDashboard = viewType === View.Type.Dashboard;
 
   const onToggleEdit = useCallback(() => {
-    sendTelemetry(TELEMETRY_EVENT_TYPE.SEARCH_WIDGET_ACTION.WIDGET_EDIT_TOGGLED, {
-      app_pathname: getPathnameWithoutId(pathname),
-      app_section: 'search-widget',
-      app_action_value: 'widget-edit-button',
-    });
-
     if (editing) {
       unsetWidgetEditing();
       setOldWidget(undefined);
     } else {
+      sendWidgetEditTelemetry();
       stopAutoRefresh();
       setWidgetEditing(widget.id);
       setOldWidget(widget);
     }
-  }, [editing, pathname, sendTelemetry, setWidgetEditing, stopAutoRefresh, unsetWidgetEditing, widget]);
+  }, [editing, sendWidgetEditTelemetry, setWidgetEditing, stopAutoRefresh, unsetWidgetEditing, widget]);
   const onCancelEdit = useCallback(() => {
-    sendTelemetry(TELEMETRY_EVENT_TYPE.SEARCH_WIDGET_ACTION.WIDGET_EDIT_CANCEL_CLICKED, {
-      app_pathname: getPathnameWithoutId(pathname),
-      app_section: 'search-widget',
-      app_action_value: 'widget-edit-cancel-button',
-    });
+    sendWidgetEditCancelTelemetry();
 
     if (oldWidget) {
       dispatch(updateWidget(id, oldWidget));
     }
 
     onToggleEdit();
-  }, [dispatch, id, oldWidget, onToggleEdit, pathname, sendTelemetry]);
+  }, [dispatch, id, oldWidget, onToggleEdit, sendWidgetEditCancelTelemetry]);
   const onRenameWidget = useCallback((newTitle: string) => dispatch(setWidgetTitle(id, newTitle)), [dispatch, id]);
   const onWidgetConfigChange = useCallback(
     async (newWidgetConfig: WidgetConfig) => {
-      sendTelemetry(TELEMETRY_EVENT_TYPE.SEARCH_WIDGET_ACTION.WIDGET_CONFIG_UPDATED, {
-        app_pathname: getPathnameWithoutId(pathname),
-        app_section: 'search-widget',
-        app_action_value: 'widget-edit-update-button',
-      });
+      sendWidgetConfigUpdateTelemetry();
 
       return dispatch(updateWidgetConfig(id, newWidgetConfig)).then(() => {});
     },
-    [dispatch, id, pathname, sendTelemetry],
+    [dispatch, id, sendWidgetConfigUpdateTelemetry],
   );
   const activeQuery = useActiveQueryId();
